i386: virt: Initialise AcpiConfiguration pointer earlier
Previously the vms->acpi_configuration pointer was setup in the "done"
notification for machine initialisation. However this is too late when
memory configuration details are provided on the command line (and thus
the memory hotplug functions are used.)

As nothing in this struct relies on values that are set between
virt_machine_state_init and virt_machine_done then we can safely
initialise this this pointer and populate the structure during
virt_machine_state_init.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>